+++ include/wfa/art_activity_bootstrap_fixture.hpp
@@ -34,6 +34,7 @@ struct NativeArtActivityBootstrapFixtureReport {
   std::string art_runtime_probe_source;
   std::string art_runtime_probe_inventory_path;
   std::string art_runtime_probe_detection_reason;
+  std::string art_runtime_probe_capability;
   bool runtime_class_resolution_succeeded = false;
   bool application_probe_attempted = false;
   bool application_probe_succeeded = false;